Move image resizing to datasets
Move image resizing from the training loop to the dataset loader,
following the convention from torchvision's built-in datasets. This
allows setting the batch size to > 1 for the HierText dataset, since it
ensures that all images yielded by the dataset have the same size.

Also improve progress reporting for tqdm by adding a progress bar.

Both of these changes are in preparation for a larger scale training
with a GPU.