Revise detection model to be more flexible
Revise the U-Net-like model to be more flexible with respect to the
number of levels and the number of channels from the output of each
level.

In the process the depth scale has been adjusted to more aggressively
downsample the inputs with a small number of feature channels in the
early layers. This will be combined with a change in training to use a
higher-resolution input, in order to help more accurately segment small
text lines that are closely spaced, without increasing inference time
too much.