Implement differentiable binarization
Implement differentiable binarization from [1][2]. This doesn't include the
entire model, just the key element of (text probabiliy mask, threshold mask,
approximate binary mask) outputs, where the approximate binary mask is computed
from the text probability mask and threshold mask using the differentiable
binarization function.

The effect of this, per analysis in [2], is to increase the gradients of the
losses around the boundary (text/not-text) value, encouraging more distinct /
less ambiguous predictions.

The threshold and approximate binary masks are only used during training. Per
the DB paper, the approximate binary mask is almost the same as the probability
mask after training, so we can skip generating the approximate binary mask and
threshold masks during inference. An `eval_only` flag has been added to the
model to additionally skip allocating the parameters needed to generate these
masks, if the model is only going to be used for inference.

Supervision is currently only applied to the text probability mask and
approximate binary mask. The paper also uses something similar to the border
mask to supervise the threshold mask.

[1] https://arxiv.org/abs/1911.08947
[2] https://arxiv.org/abs/2202.10304