Compute word-level metrics during test loop
Compute word-level metrics (precision, recall, merged and split boxes) for each
image processed during the test loop, and return a dict containing the means of
those metrics at the end. These metrics are now formatted and printed at the
end of each validation loop. This provides a way to see progress of training
at the word level as well as the existing pixel-level progress.