Rework upscaled tensor size trimming to simplify ONNX model graph
The ONNX output sub-graph generated for adjusting an upscaled tensor to
match the input coming across was unexpectedly complex and various
warnings were produced during export. Since the `x_to_upscale` tensor is
half the height/width of the `x` tensor and is being ConvTranspose2d'd
with a 3x3 kernel and stride 2, the upscaled result is actually either
the same size or 1 pixel wider/taller than `x`. Hence we can trim it to
match with a simpler slice operation. This fixes the warnings and makes
the ONNX model structure much simpler.