Cache text line images on disk
When the `HierTextRecognition` dataset is being read with a `DataLoader` that
has `shuffle=True`, the shuffling defeats the LRU cache used when reading full
images. Shuffling the dataset is important for good training results, so change
the caching in the dataset to instead save cached text line images to disk,
which are cheaper to read than the full image.