Round up image widths / text lengths to multiple of 250 when creating batches
When training with PyTorch 2 on an NVidia A10 I encountered a problem where the
Python process memory usage would steadily increase during training and fail
with an OOM after 1.5 epochs or so. Rounding up the padded image size to a
multiple of some large unit, I chose 250 here, reduces the number of distinct
tensor sizes during training and so far seems to have alleviated the problem.