Speed up training using automatic mixed-precision
Use `torch.autocast` with bfloat16 for a modest training speed up - epochs
reduced from 15m to 11m30s on an NVidia A10. The bfloat16 dtype is used as
the model eventually started to produce NaN losses when using float16. I got
an error when using bf16 with the GRU layer, so that region disables autocast.