Optimize character error rate calculation
It turns out that indexing into tensors in a loop in `decode_text` and
`ctc_greedy_decode_text` is extremely slow, compared to indexing into a
`list[int]`, and is several times worse if the tensor is on the GPU. Solve this
by converting the `[batch, seq]` tensors to `list[list[int]]` in one step,
before looping over items in the batch. The first version of CER stats increased
validation set evaluation times from ~34s to 12 minutes (!!) on an EC2 server
with A10 GPU. This brings it back down to ~40s.