Add `--trim` option to layout scraper
This "trims" the text layout saved to the JSON file so that it only includes
text that appears in the viewport. This makes the output align with the text
that appears in the screenshot, as well as more representative of a typical
document image that might be OCRed, which would usually be one screenshot or
page of a document.