Take into account input padding when decoding outputs from model
When decoding outputs from a model to preview or compute the char error rate,
the part of the output that corresponds to the padding region of the input needs
to be ignored. The loss function only takes into consideration the un-padded
part of the input.

This fixes an issue where predictions had spurious extra characters on the end
and the calculated char error rate was higher than it should have been.