Change CTC loss reduction
Change the way CTC losses for each output step are reduced. Instead of using
`reduction="mean"` which averages over the target length and then taking the
mean over the batch, use `reduction="sum"` and then divide by the batch size.
This produces larger gradients which resulted in moderately faster training and
larger/more readable loss values. The small-ness of the gradients previously was
compensated for by the Adam optimizer, so the benefit here is not as dramatic as
if vanilla SGD was used.